% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_models}
\alias{available_models}
\title{List available shorthand models}
\usage{
available_models()
}
\value{
Character vector of available model shorthands
}
\description{
List available shorthand models
}
\examples{
available_models()
}
