% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear model cache}
\usage{
clear_cache(confirm = TRUE)
}
\arguments{
\item{confirm}{Logical, if TRUE will prompt for confirmation}
}
\value{
Logical indicating success
}
\description{
Clear model cache
}
\examples{
\dontrun{
clear_cache(confirm = FALSE)
}
}
