% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{create_standard_symlinks}
\alias{create_standard_symlinks}
\title{Create standard symlinks for model files}
\usage{
create_standard_symlinks(model_dir)
}
\arguments{
\item{model_dir}{Path to model directory}
}
\description{
Create standard symlinks for model files
}
\keyword{internal}
