% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cuda_available}
\alias{cuda_available}
\title{Check if CUDA is available}
\usage{
cuda_available()
}
\value{
Logical indicating if CUDA support is available
}
\description{
Checks whether the package was installed with CUDA support and whether
the CUDA provider libraries are present.
}
\examples{
cuda_available()
}
