% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{detect_model_type}
\alias{detect_model_type}
\title{Detect model type from files in directory}
\usage{
detect_model_type(model_dir)
}
\arguments{
\item{model_dir}{Path to model directory}
}
\value{
String indicating model type
}
\description{
Detect model type from files in directory
}
\keyword{internal}
