% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_bytes}
\alias{format_bytes}
\title{Format file size for display}
\usage{
format_bytes(bytes)
}
\arguments{
\item{bytes}{Number of bytes}
}
\value{
Formatted string (e.g., "1.5 MB")
}
\description{
Format file size for display
}
\keyword{internal}
