% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model_config}
\alias{get_model_config}
\title{Get model configuration for sherpa-onnx C API}
\usage{
get_model_config(model_info)
}
\arguments{
\item{model_info}{Model info list from resolve_model()}
}
\value{
List with model configuration parameters
}
\description{
Get model configuration for sherpa-onnx C API
}
\keyword{internal}
