% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_model_dir}
\alias{is_valid_model_dir}
\title{Check if a path is a valid model directory}
\usage{
is_valid_model_dir(path)
}
\arguments{
\item{path}{Path to check}
}
\value{
Logical indicating if path contains model files
}
\description{
Check if a path is a valid model directory
}
\keyword{internal}
