% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{new_sherpa_transcription}
\alias{new_sherpa_transcription}
\title{Create a sherpa_transcription object}
\usage{
new_sherpa_transcription(result_list, model_info)
}
\arguments{
\item{result_list}{List containing transcription results from C++ layer}

\item{model_info}{Model metadata from resolve_model()}
}
\value{
A sherpa_transcription object (S3 class inheriting from list)
}
\description{
Create a sherpa_transcription object
}
\keyword{internal}
