% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{new_sherpa_vad_result}
\alias{new_sherpa_vad_result}
\title{Create a sherpa_vad_result object}
\usage{
new_sherpa_vad_result(segments, num_segments, sample_rate, source_file)
}
\arguments{
\item{segments}{List of segment objects from extract_vad_segments_()}

\item{num_segments}{Number of segments}

\item{sample_rate}{Audio sample rate}

\item{source_file}{Original WAV file path}
}
\value{
A sherpa_vad_result object
}
\description{
Create a sherpa_vad_result object
}
\keyword{internal}
