% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{print.sherpa_vad_result}
\alias{print.sherpa_vad_result}
\title{Print method for sherpa_vad_result}
\usage{
\method{print}{sherpa_vad_result}(x, ...)
}
\arguments{
\item{x}{A sherpa_vad_result object}

\item{...}{Additional arguments (ignored)}
}
\value{
The object invisibly
}
\description{
Print method for sherpa_vad_result
}
