% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_wav}
\alias{read_wav}
\title{Read a WAV file}
\usage{
read_wav(wav_path)
}
\arguments{
\item{wav_path}{Path to WAV file}
}
\value{
List with samples, sample_rate, and num_samples
}
\description{
Read a WAV file
}
\keyword{internal}
