% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{summary.sherpa_vad_result}
\alias{summary.sherpa_vad_result}
\title{Summary method for sherpa_vad_result}
\usage{
\method{summary}{sherpa_vad_result}(object, ...)
}
\arguments{
\item{object}{A sherpa_vad_result object}

\item{...}{Additional arguments (ignored)}
}
\value{
The object invisibly
}
\description{
Summary method for sherpa_vad_result
}
