% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{vad_segment_samples}
\alias{vad_segment_samples}
\title{Extract audio samples from a VAD segment}
\usage{
vad_segment_samples(vad_result, segment_index)
}
\arguments{
\item{vad_result}{A sherpa_vad_result object}

\item{segment_index}{Index of segment to extract (1-based)}
}
\value{
Numeric vector of audio samples
}
\description{
Extract audio samples from a VAD segment
}
\examples{
\dontrun{
result <- vad("recording.wav")
# Get samples from first segment
samples <- vad_segment_samples(result, 1)
}

}
